Extract `decode_body`, adjust tests
In this commit I've made each field obligatory, removing the `Option`, which
jives with our current understanding of the agent's behavior. This commit
introduces a separate decoding function with quickcheck test to verify, which
makes testing just the decoding step much more direct that previously. Some of
the insertion calls are adjusted to avoid inefficiencies found in `try_insert`,
partially addressed in #8331

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>